deepseek-mini/requirements.txt
torch>=2.0.0
torchvision
transformers>=4.30.0
tokenizers>=0.13.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
tqdm>=4.65.0
pyyaml>=6.0
PyQt5>=5.15.0
PyQt5-Qt5>=5.15.2
PyQt5-sip>=12.11.0
qdarkstyle>=3.1
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
jupyter>=1.0.0
ipython>=8.12.0
requests>=2.28.0
websockets>=11.0.0
accelerate>=0.18.0
datasets>=2.12.0
safetensors>=0.3.0
einops>=0.6.0
wandb>=0.15.0
tensorboard>=2.13.0